FINAL WORKING VERSION

+++ timestamp.c
@@ -3,7 +3,7 @@
 
 // Dr. Butler's class
 
-float timestamp()
+double timestamp()
 {
   struct timeval timer;
